fix: refactor default mapping structure
