+++ packages/react/src/themes/default-theme/default-alias-tokens.ts
@@ -1,0 +1,11 @@
+import { AliasTokenMap } from '../tokens';
+
+export const defaultAliasTokens: AliasTokenMap = {
+    'focus-border-color': 'color-brand-50',
+    'focus-box-shadow-color': 'color-brand-20',
+    'focus-box-shadow-inset-color': 'color-brand-20',
+    'focus-border-box-shadow-color-1': 'color-brand-50',
+    'focus-border-box-shadow-color-2': 'color-brand-20',
+    'focus-border-box-shadow-inset-color-1': 'color-brand-20',
+    'focus-border-box-shadow-inset-color-2': 'color-brand-50',
+};
