+++ packages/react/src/themes/default-theme/default-ref-tokens.ts
@@ -1,66 +1,10 @@
-import { ColorUtilityTokenMap } from '../tokens/ref-tokens/color-utility-tokens';
-import { PaletteTokenMap } from '../tokens/ref-tokens/palette-tokens';
-import { TextAttributeTokenMap } from '../tokens/ref-tokens/text-attributes-tokens';
+import { RefTokenMap } from '../tokens';
+import { defaultColorUtilityTokens } from './default-ref-tokens/default-color-utility-tokens';
+import { defaultPaletteTokens } from './default-ref-tokens/default-palette-tokens';
+import { defaultTextAttributeTokens } from './default-ref-tokens/default-text-attributes-tokens';
 
-export const defaultPaletteTokens: PaletteTokenMap = {
-    'color-brand-05': '#E0F0F9',
-    'color-brand-20': '#84C6EA',
-    'color-brand-50': '#006296',
-    'color-brand-70': '#003A5A',
-    'color-brand-80': '#012639',
-    'color-accent-20': '#F9B6B2',
-    'color-accent-50': '#EF483E',
-    'color-accent-70': '#D41F14',
-    'color-white': '#FFFFFF',
-    'color-black': '#000000',
-    'color-neutral-02': '#FAFAFA',
-    'color-neutral-05': '#F1F2F2',
-    'color-neutral-15': '#DBDEE1',
-    'color-neutral-30': '#B7BBC2',
-    'color-neutral-50': '#878F9A',
-    'color-neutral-65': '#60666E',
-    'color-neutral-90': '#1B1C1E',
-    'color-alert-02': '#FDF7F6',
-    'color-alert-05': '#FAEAE9',
-    'color-alert-20': '#F99D99',
-    'color-alert-50': '#CD2C23',
-    'color-alert-70': '#7B1A15',
-    'color-informative-02': '#F9F7FB',
-    'color-informative-05': '#E0F0F9',
-    'color-informative-20': '#84C6EA',
-    'color-informative-50': '#006296',
-    'color-informative-70': '#003A5A',
-    'color-success-02': '#F6FCF8',
-    'color-success-05': '#E1F7EA',
-    'color-success-20': '#8ADDA9',
-    'color-success-50': '#008533',
-    'color-success-70': '#004F1E',
-    'color-warning-02': '#FFFBF5',
-    'color-warning-05': '#FFF7E5',
-    'color-warning-20': '#FFDD99',
-    'color-warning-50': '#F5A200',
-    'color-warning-70': '#9E6900',
-    'color-warning-80': '#664400',
-    'color-discovery-02': '#F9F7FB',
-    'color-discovery-05': '#EFEAF6',
-    'color-discovery-20': '#CFC1E3',
-    'color-discovery-50': '#602FA0',
-    'color-discovery-70': '#3A1C60',
-};
-
-export const defaultColorUtilityTokens: ColorUtilityTokenMap = {
-    'transparent-10': 'rgba(0, 0, 0, 0.1)',
-    'transparent-20': 'rgba(0, 0, 0, 0.2)',
-    'transparent-30': 'rgba(0, 0, 0, 0.3)',
-    'transparent-40': 'rgba(0, 0, 0, 0.4)',
-    'transparent-50': 'rgba(0, 0, 0, 0.5)',
-    'transparent-60': 'rgba(0, 0, 0, 0.6)',
-    'transparent-70': 'rgba(0, 0, 0, 0.7)',
-    'transparent-80': 'rgba(0, 0, 0, 0.8)',
-    'transparent-90': 'rgba(0, 0, 0, 0.9)',
-    'transparent-100': 'transparent',
-};
-
-export const defaultTextAttributeTokens: TextAttributeTokenMap = {
-    'font-weight-bold': 'bold',
+export const defaultRefTokens: RefTokenMap = {
+    ...defaultPaletteTokens,
+    ...defaultColorUtilityTokens,
+    ...defaultTextAttributeTokens,
 };
