+++ packages/react/src/themes/default-theme/default-ref-tokens/default-color-utility-tokens.ts
@@ -1,0 +1,14 @@
+import { ColorUtilityTokenMap } from '../../tokens';
+
+export const defaultColorUtilityTokens: ColorUtilityTokenMap = {
+    'transparent-10': 'rgba(0, 0, 0, 0.1)',
+    'transparent-20': 'rgba(0, 0, 0, 0.2)',
+    'transparent-30': 'rgba(0, 0, 0, 0.3)',
+    'transparent-40': 'rgba(0, 0, 0, 0.4)',
+    'transparent-50': 'rgba(0, 0, 0, 0.5)',
+    'transparent-60': 'rgba(0, 0, 0, 0.6)',
+    'transparent-70': 'rgba(0, 0, 0, 0.7)',
+    'transparent-80': 'rgba(0, 0, 0, 0.8)',
+    'transparent-90': 'rgba(0, 0, 0, 0.9)',
+    'transparent-100': 'transparent',
+};
