+++ packages/react/src/themes/default-theme/default-ref-tokens/default-text-attributes-tokens.ts
@@ -1,0 +1,5 @@
+import { TextAttributeTokenMap } from '../../tokens';
+
+export const defaultTextAttributeTokens: TextAttributeTokenMap = {
+    'font-weight-bold': 'bold',
+};
